Implemented a rudimentary existential variable value selection mechanism.
We select for existential variables those values that induce goal atoms with minimum sums of hmax
values. That is, if an existential variable Z participates in 2 goal atoms p(Z) and q (Z, c), we
will
prioritize the possible values z of Z such that hmax(p(z)) + hmax(q(z,c)) is minimum, i.e. such that
the
induced atoms are achieved the earliest possible