redefine BrFS in accordance with Russell&Norvig:
The only modification is that the check for whether a state is a goal
or not is done right after the creation of the state, instead of upon expansion.
On a problem that has a solution at depth 'd', this avoids the worst-case expansion
of all the $b^d$ nodes of the last (deepest) layer (where b is the branching
factor).